remove common folder and component
